Split class into files
