+++ conjur/argument_parser/_host_parser.py
@@ -1,0 +1,58 @@
+import argparse
+from conjur.argument_parser.parser_utils import command_description, command_epilog, formatter, title_formatter
+
+
+class HostParser:
+    def add_host_parser(self):
+        """
+        Method adds host parser functionality to parser
+        """
+        host_subparser = self._init_host_parser()
+        host_subparsers = host_subparser.add_subparsers(dest='action', title=title_formatter("Subcommands"))
+        HostParser._host_rotate_api_key_parser(host_subparsers)
+        HostParser._add_host_options(host_subparser)
+
+        return self
+
+    def _init_host_parser(self):
+        host_name = 'host - Manage hosts'
+        host_usage = 'conjur [global options] host <subcommand> [options] [args]'
+        host_subparser = self.resource_subparsers.add_parser('host',
+                                                             help='Manage hosts',
+                                                             description=command_description(host_name,
+                                                                                             host_usage),
+                                                             epilog=command_epilog(
+                                                                 'conjur host rotate-api-key -i my_apps/myVM\t\t'
+                                                                 'Rotates the API key for host myVM',
+                                                                 command='host',
+                                                                 subcommands=['change-password']),
+                                                             usage=argparse.SUPPRESS,
+                                                             add_help=False,
+                                                             formatter_class=formatter)
+
+        return host_subparser
+
+    @staticmethod
+    def _host_rotate_api_key_parser(sub_parser):
+        host_rotate_api_key_name = 'rotate-api-key - Rotate a host\'s API key'
+        host_rotate_api_key_usage = 'conjur [global options] host rotate-api-key [options] [args]'
+        host_rotate_api_key_parser = sub_parser.add_parser('rotate-api-key',
+                                                                help='Rotate a host\'s API key',
+                                                                description=command_description(
+                                                                    host_rotate_api_key_name,
+                                                                    host_rotate_api_key_usage),
+                                                                epilog=command_epilog(
+                                                                    'conjur host rotate-api-key -i my_apps/myVM\t\t'
+                                                                    'Rotates the API key for host myVM'),
+                                                                usage=argparse.SUPPRESS,
+                                                                add_help=False,
+                                                                formatter_class=formatter)
+        host_rotate_api_key = host_rotate_api_key_parser.add_argument_group(title=title_formatter("Options"))
+        host_rotate_api_key.add_argument('-i', '--id',
+                                         help='Provide host identifier for which you want to rotate the API key')
+        host_rotate_api_key.add_argument('-h', '--help', action='help', help='Display help screen and exit')
+
+    @staticmethod
+    def _add_host_options(host_subparser):
+        host_options = host_subparser.add_argument_group(title=title_formatter("Options"))
+        host_options.add_argument('-h', '--help', action='help', help='Display help screen and exit')
