+++ conjur/argument_parser/_screen_options_parser.py
@@ -1,0 +1,25 @@
+from conjur.version import __version__
+from conjur.argument_parser.parser_utils import conjur_copyright
+
+
+class ScreenOptionsParser:
+    def add_main_screen_options(self):
+        """
+        Method adds main screen options functionality to parser
+        """
+        global_optional = self.parser.add_argument_group("Global options")
+        global_optional.add_argument('-h', '--help', action='help', help="Display help list")
+        global_optional.add_argument('-v', '--version', action='version',
+                                     help="Display version number",
+                                     version='Conjur CLI version ' + __version__ + "\n"
+                                             + conjur_copyright())
+
+        global_optional.add_argument('-d', '--debug',
+                                     help='Enable debugging output',
+                                     action='store_true')
+
+        global_optional.add_argument('--insecure',
+                                     help='Skip verification of server certificate (not recommended for production).\nThis makes your system vulnerable to security attacks!\n',
+                                     dest='ssl_verify',
+                                     action='store_false')
+        return self
